feat(web): relocate layout components
