+++ Dashboard/components/layout/navigation/Header.tsx
@@ -1,0 +1,93 @@
+import { Burger, Container, createStyles, Group, Header } from "@mantine/core";
+import { useDisclosure } from "@mantine/hooks";
+import NextButton from "./NextButton";
+
+const useStyles = createStyles((theme) => ({
+  header: {
+    display: "flex",
+    justifyContent: "space-between",
+    alignItems: "center",
+    height: "100%",
+  },
+
+  links: {
+    [theme.fn.smallerThan("xs")]: {
+      display: "none",
+    },
+  },
+
+  burger: {
+    [theme.fn.largerThan("xs")]: {
+      display: "none",
+    },
+  },
+
+  link: {
+    display: "block",
+    lineHeight: 1,
+    padding: "8px 12px",
+    borderRadius: theme.radius.sm,
+    textDecoration: "none",
+    color:
+      theme.colorScheme === "dark"
+        ? theme.colors.dark[0]
+        : theme.colors.gray[7],
+    fontSize: theme.fontSizes.sm,
+    fontWeight: 500,
+
+    "&:hover": {
+      backgroundColor:
+        theme.colorScheme === "dark"
+          ? theme.colors.dark[6]
+          : theme.colors.gray[0],
+    },
+  },
+
+  linkActive: {
+    "&, &:hover": {
+      backgroundColor: theme.fn.variant({
+        variant: "light",
+        color: theme.primaryColor,
+      }).background,
+      color: theme.fn.variant({
+        variant: "light",
+        color: theme.primaryColor,
+      }).color,
+    },
+  },
+}));
+
+interface HeaderSimpleProps {
+  links: { link: string; label: string }[];
+}
+
+export function HeaderSimple({ links }: HeaderSimpleProps) {
+  const [opened, { toggle }] = useDisclosure(false);
+  const { classes } = useStyles();
+
+  const items = links.map((link) => (
+    <NextButton
+      color="blue"
+      key={`${link.label}-`}
+      href={link.link}
+      title={link.label}
+    />
+  ));
+
+  return (
+    <Header height={60} mb={120}>
+      <Container className={classes.header}>
+        KU Blockchain Explorer
+        <Group spacing={5} className={classes.links}>
+          {items}
+        </Group>
+        <Burger
+          opened={opened}
+          onClick={toggle}
+          className={classes.burger}
+          size="sm"
+        />
+      </Container>
+    </Header>
+  );
+}
